Change xml4_document to render using full row
Instead of pulling out doc or value from the view row, use the entire
row to render xml fragments. This allows templates to work with both a
view row’s document and value at the same time. And, it’s easier to
understand because a template row now lines up exactly with a view row.