Combine section type and number for long desc
In both the offerings and sections list functions, create a variable in
the row that combines the section type and number. It is not possible
to pick out these values from code_info.components in Dust, since there
appears to be no way to access specific members of an array.