Declare Current Semester for Cont Ed Courses
Continuing Education courses are specified to be in a semester named
“Continuing Education” for D2L. Because there is no “Continuing
Education” semester in Ares, these courses end up in the furthermost
declared semester, instead. The Copyright Office would prefer they be
placed in approximately the “right” semester.

The incoming data for Destiny One does not indicate when a course
occurs. Modify the ares_courses function to output the current semester
(based on the current month and year) as the semester for Continuing
Education courses. Although this will put past CE courses in the wrong
semester, this is relying on the behaviour that already created courses
in Ares are not updated by changes in the daily data feeds. So, this
should effectively cause newly appearing courses to be placed in the
current semester, without adversely impacting past courses.